+++ packages/ricos-editor/web/src/utils/editorCommandsTestsUtil.ts
@@ -483,20 +483,24 @@ const image = {
 
 const link = {
   type: RICOS_LINK_TYPE,
-  data1: LinkData.fromJSON({ url: 'www.wix.com', rel: 'nofollow', target: '_self' }),
+  data1: LinkData.fromJSON({
+    link: {
+      url: 'www.wix.com',
+      rel: { nofollow: true },
+      target: 'SELF',
+    },
+  }),
   selection1: selection,
-  data2: LinkData.fromJSON({ url: 'www.sport5.co.il', rel: 'sponsored', target: '_self' }),
+  data2: LinkData.fromJSON({
+    link: {
+      url: 'www.sport5.co.il',
+      rel: { sponsored: true },
+      target: 'SELF',
+    },
+  }),
   selection2: selectionCollapsed,
-  expectedData1: {
-    url: 'www.wix.com',
-    rel: 'nofollow',
-    target: '_self',
-  },
-  expectedData2: {
-    url: 'www.sport5.co.il',
-    rel: 'sponsored',
-    target: '_self',
-  },
+  expectedData1: { url: 'www.wix.com', rel: 'nofollow', target: '_self' },
+  expectedData2: { url: 'www.sport5.co.il', rel: 'sponsored', target: '_self' },
 };
 
 const mention = {